requirement.txt
seaborn
matplotlib
pandas
numpy
wordcloud
streamlit
pyyaml
